Add Redis store and refactor code
This commit adds support for a Redis store in the `rex.go` file. It initializes the Redis store with the specified options and updates the code to use the new store. Additionally, the `SubscribeToSensorUpdates` function has been modified to extract sensor keys from the compiled rules and handle data in the expected format. Some minor code refactoring has also been done.